Move compile_circuit to separate module
